soap/client/requirements.txt
pynecone-io
zeep